refactor: 마이페이지 리팩토링(타입, 데이터 호출 분리, listItem를 IndexNav 컴포넌트로 분리)
